requirements.txt
# CareerGuideAI Dependencies
# Core Python packages for the career guidance system

# Web Application Dependencies
Flask>=2.3.0
Werkzeug>=2.3.0
gunicorn
reportlab
python-dotenv
Flask-Limiter
pydantic>=2.0
Flask-SQLAlchemy
Flask-Login
Flask-Caching
celery
redis
psycopg2-binary

# Optional dependencies for enhanced functionality (uncomment if needed):
# requests>=2.28.0  # For API integrations
# pandas>=1.5.0     # For data analysis
# matplotlib>=3.6.0 # For trend visualization
# streamlit>=1.24.0 # For web interface gunicorn 
 
 